Working with SQL
I was required to track a criminal. the only information i had was 3 reports from the police investigation, the day of the incident, camera feed, and the location.

In the end, i was able to  fine the culprit, his partner, the flight he used and the country he seemingly escaped to.